+++ src/Controller/menu/Playable.java
@@ -1,13 +1,9 @@
 package Controller.menu;
 
-import Model.Table.Cell;
-
-import java.util.ArrayList;
-
 public interface Playable {
     void setProbability(int... probability);
     void endTurn();
-    int[][] play();
+    float[][] play();
     int getTurn();
 
 }
